Bluetooth protocol and structure updated

+++ app/src/main/java/com/example/ksala/healthproject/DataMessage.java
@@ -4,9 +4,9 @@ package com.example.ksala.healthproject;
  * Created by kevin on 06/12/2015.
  */
 public class DataMessage {
-    public int x, y;
+    public double x, y;
 
-    public DataMessage(int x, int y) {
+    public DataMessage(double x, double y) {
         this.x = x;
         this.y = y;
     }
